fix quicksort

+++ include/sort.h
@@ -1,12 +1,14 @@
 #ifndef _SPCAL_SORT_
 #define _SPCAL_SORT_
 
+#include <stdlib.h>
+
 struct argsort {
   double value;
-  int index;
+  size_t index;
 };
 
-void mergesort_argsort(struct argsort *x, int n);
-void quicksort_argsort(struct argsort *x, int n);
+void mergesort_argsort(struct argsort *x, size_t n);
+void quicksort_argsort(struct argsort *x, size_t n);
 
 #endif
